OneDrive/partyoria_landingpage/partyoria-landing-page/partyoria-event-hub-main/backend/venues/serializers.py
from rest_framework import serializers
from .models import VenueDetails

class VenueDetailsSerializer(serializers.ModelSerializer):
    class Meta:
        model = VenueDetails
        fields = ['id', 'venue_name', 'location', 'capacity', 'price_range', 'image_url', 'description', 'created_at']