OneDrive/partyoria_landingpage/partyoria-landing-page/partyoria-event-hub-main/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.9